Fix http-codec failing occasionally during chunked decoding
It is not guaranteed that the data we have received so far will include both the chunk-size and its terminator, so failing in that case is wrong. Instead, we should wait until we get more data and check again. This commit makes it so that decoding only fails if we still don't find a terminator with a chunk-size field of >= 32 hex digits, which should never happen under normal circumstances.

Detailed example of when this can occur:

- Data received so far: '4\r\nTest\r\n13\r'
- First chunk decoded: len='4', data='Test'
- Data remaining to decode: '13\r'
- Attempt to decode next chunk: len='1', term='3\r'; term ~= '\r\n' (without this commit, the code would error here with a failed assert)
- New data received: '\nFixed in this case\r\n'
- Data remaining to decode: '13\r\nFixed in this case\r\n'
- Re-attempt to decode chunk: len='13', data='Fixed in this case'