Add helper script to determine the required Luvi version
This can be used by the CI to determine which Luvi version to integrate in the build, without having to specify it anywhere else than in the package metadata (where it's already present).

I don't think this is an ideal solution, but I haven't found any that would be less messy. Since the script is "hidden" in the CI folder, it likely won't cause any harm.