Tests: Add basic test for the lit help command
This is also the default text that's displayed when no arguments were passed, so both cases are covered.

The purpose of this isn't do establish a proper and generic test suite, which is way too painful without a proper framework, but rather to make sure I can refactor the CLI code without breaking things.