fix: invalid usage of uv.run() in command runner
We have been calling `uv.run()` from inside of a
libuv callback, which is not allowed. This moves
the original call to `uv.run()` inside of the main
xpcall, which is where it is necessary.

This allows the first call to `uv.run()` to be
completely finished before we attempt to exit and
call `uv.run()` again.